+++ packages/core/src/binders/attribute.binder.spec.ts
@@ -1,12 +1,12 @@
 import { Riba } from "../riba";
 
-import { starBinder } from "./attribute.binder";
+import { attributeBinder } from "./attribute.binder";
 
 import { dotAdapter } from "../adapters/dot.adapter";
 
 const riba = new Riba();
 riba.module.adapter.regist(dotAdapter);
-riba.module.binder.regist(starBinder);
+riba.module.binder.regist(attributeBinder);
 
 describe("riba.binders", () => {
   let element: HTMLDivElement;